# Optimize DATE_TRUNC, DATE_ADD and DATE_DIFF functions for DATE and TIMESTAMP types