<!-- Thank you for submitting a pull request! Find more information in our development guide at https://github.com/trinodb/trino/blob/master/.github/DEVELOPMENT.md and contact us on #dev in Slack. -->
<!-- Provide an overview of the PR for maintainers and reviewers. -->
## Description
We have observed poor performance when using the `DATE_TRUNC` function. Profiling revealed that the function spends most of its time resolving date time units. The main bottleneck is in the `io.trino.operator.scalar.DateTimeFunctions#getDateField` method. This method contains 2 string allocations and a hash lookup, which takes a noticeable amount of CPU time.

![before](https://github.com/trinodb/trino/assets/11261786/f7558b72-7b68-4873-a797-1eea66ec09d5)

The new approach doesn't create new strings and works directly with bytes. This is possible because the date time unit can only use ASCII symbols. Also, the hash lookup has been replaced with a simple loop, which works quite well with small amounts of values.

![after](https://github.com/trinodb/trino/assets/11261786/b3be55e8-5774-4298-8bdf-6157b72b3559)

Here is the benchmark results. All the tests were performed on Apple Silicon M1 MAX, OpenJDK 64-Bit Server VM, 22.0.1+8.
```
Benchmark                                        (unitString)  Mode  Cnt    Score    Error  Units
BenchmarkDateTimeFunctions.getTimestampFieldNew   millisecond  avgt   20    6.578 ±  0.222  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew   MILLISECOND  avgt   20    6.440 ±  0.043  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld   millisecond  avgt   20   24.328 ±  2.810  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld   MILLISECOND  avgt   20   64.824 ±  1.651  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew        second  avgt   20    5.303 ±  0.083  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew        SECOND  avgt   20    5.408 ±  0.360  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld        second  avgt   20   26.958 ±  0.648  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld        SECOND  avgt   20   84.577 ± 11.841  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew        minute  avgt   20    7.124 ±  0.151  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew        MINUTE  avgt   20    7.041 ±  0.148  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld        minute  avgt   20   27.429 ±  0.475  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld        MINUTE  avgt   20   73.585 ± 11.993  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew          hour  avgt   20    5.341 ±  0.377  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew          HOUR  avgt   20    4.897 ±  0.064  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld          hour  avgt   20   33.435 ±  2.004  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld          HOUR  avgt   20   56.233 ±  4.317  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew           day  avgt   20    5.061 ±  0.126  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew           DAY  avgt   20    5.022 ±  0.033  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld           day  avgt   20   28.863 ±  1.056  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld           DAY  avgt   20   76.999 ±  8.112  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew          week  avgt   20    8.032 ±  0.144  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew          WEEK  avgt   20    7.909 ±  0.040  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld          week  avgt   20   28.985 ±  1.604  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld          WEEK  avgt   20   84.112 ±  8.137  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew         month  avgt   20    6.530 ±  0.069  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew         MONTH  avgt   20    6.568 ±  0.094  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld         month  avgt   20   29.953 ±  1.834  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld         MONTH  avgt   20   83.186 ± 10.773  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew       quarter  avgt   20   22.135 ±  0.877  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew       QUARTER  avgt   20   22.200 ±  0.799  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld       quarter  avgt   20   78.707 ±  5.599  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld       QUARTER  avgt   20  115.353 ±  8.914  ns/op

BenchmarkDateTimeFunctions.getTimestampFieldNew          year  avgt   20   11.206 ±  0.370  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldNew          YEAR  avgt   20   11.020 ±  0.266  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld          year  avgt   20   29.169 ±  0.842  ns/op
BenchmarkDateTimeFunctions.getTimestampFieldOld          YEAR  avgt   20   67.568 ± 11.769  ns/op
```
The benchmark shows that the new lookup implementation is 4-10 times faster. Interestingly, the new algorithm also removes the disrepancy between lowercase and mixed-case unit representation. In the current Trino version, `DATE_TRUNC('month', date)` works faster than `DATE_TRUNC('MONTH', date)`.

To test a close-to-real-world scenario, the following query was used. This cross join produces ~1.8 billion rows.
```
explain analyze select date_trunc('month', a.i_rec_start_date) from tpcds.sf1.item a, tpcds.sf10.item b;
```

`EXPLAIN` shows that the `DATE_TRUNC` CPU usage has been significantly reduced.
Before optimization
```
Project[]                                                                                                                                                                             
│   Layout: [expr:date]                                                                                                                                                               
│   Estimates: {rows: ? (?), cpu: ?, memory: 0B, network: 0B}                                                                                                                         
│   CPU: 1.28m (96.96%), Scheduled: 1.32m (96.39%), Blocked: 0.00ns (0.00%), Output: 1836000000 rows (8.55GB)                                                                         
│   Input avg.: 459000000.00 rows, Input std.dev.: 173.21%                                                                                                                            
│   expr := date_trunc(varchar(5) 'month', i_rec_start_date)   
```
After optimization
```
Project[]                                                                                                                                                                               
│   Layout: [expr:date]                                                                                                                                                                 
│   Estimates: {rows: ? (?), cpu: ?, memory: 0B, network: 0B}                                                                                                                           
│   CPU: 41.59s (95.67%), Scheduled: 44.83s (95.75%), Blocked: 0.00ns (0.00%), Output: 1836000000 rows (8.55GB)                                                                         
│   Input avg.: 459000000.00 rows, Input std.dev.: 173.21%                                                                                                                              
│   expr := date_trunc(varchar(5) 'month', i_rec_start_date)
```

<!-- Provide details that would help an engineer who is unfamiliar with this part of the code. -->
## Additional context and related issues
**Note 1**
The same performance issue can occur with the `at_timezone` function, which uses even more complex timezone resolution logic for each call. Ideally, if a constant literal is used in a query, it should only be resolved once. 
For example, instead of calling `DateTimeFunctions#truncateDate(Slice unit, long date)`, Trino should call `DateTimeFunctions#truncateDate(DateTimeField unit, long date)`.

**Note 2**
This optimization covers only DATE and TIMESTAMP types. If the maintainers are OK with the proposed solution, the optimization for TIME type can also be implemented using the same approach.

<!-- Mark the appropriate option with an (x). Propose a release note if you can. -->
## Release notes

( ) This is not user-visible or is docs only, and no release notes are required.
( ) Release notes are required. Please propose a release note for me.
(x) Release notes are required, with the following suggested text:

```markdown
# Section
* Improved performance for DATE_TRUNC, DATE_ADD and DATE_DIFF functions
```