@alekkol 2024-06-09
Hi, @martint @wendigo ! Would be great to have you feedback here. With this patch, we have observed a 10-20% reduction in CPU usage on our Trino cluster workloads (we use DATE_TRUNC a lot). It would be great if this PR could be included in the Trino mainline so that everyone can benefit from this optimization.