@wendigo 2024-06-17
In the future, we should probably support function currying so that constant arguments are resolved only once.